need ak to write activation wrapper

+++ src/convMatrix.cpp
@@ -1,5 +1,5 @@
 #include "../include/convMatrix.hpp"
-void convMatrix::_init(float* data, int64 D0, int64 D1, int64 D2, int64 D3) {
+void convMatrix::_init(float* data, int D0, int D1, int D2, int D3) {
   _data = data;
   _D0 = D0;
   _D1 = D1;
@@ -12,12 +12,12 @@ convMatrix::convMatrix() {
   _init(NULL, 0, 0, 0, 0);
 }
 
-convMatrix::convMatrix(int64 D0, int64 D1, int64 D2, int64 D3) {
+convMatrix::convMatrix(int D0, int D1, int D2, int D3) {
   _init(NULL, D0, D1, D2, D3);
   this->_data = _numElements > 0 ? new float[this->_numElements] : NULL;
 }
 
-convMatrix::convMatrix(float* data, int64 D0, int64 D1, int64 D2, int64 D3) {
+convMatrix::convMatrix(float* data, int D0, int D1, int D2, int D3) {
   _init(data, D0, D1, D2, D3);
 }
 
@@ -26,10 +26,10 @@ void convMatrix::assign(float* data) {
 }
 
 void convMatrix::print_data() {
-  for (int64 i = 0; i < _D0; i++) {
-    for (int64 j = 0; j < _D1; j++) {
-      for (int64 m = 0; m < _D2; m++) {
-        for (int64 n = 0; n < _D3; n++) {
+  for (int i = 0; i < _D0; i++) {
+    for (int j = 0; j < _D1; j++) {
+      for (int m = 0; m < _D2; m++) {
+        for (int n = 0; n < _D3; n++) {
           std::cout << getCell(i,j,m,n) << " ";
         }
         std::cout << "\n";
@@ -39,28 +39,28 @@ void convMatrix::print_data() {
   }
 }
 
-void convMatrix::convolve(convMatrix &filter, int64 stride, bool samePadding,  convMatrix &target) {
-  int64 batch = getDim(0);
-  int64 in_height = getDim(1);
-  int64 in_width = getDim(2);
-  int64 in_channels = getDim(3);
-  int64 filter_height = filter.getDim(0);
-  int64 filter_width = filter.getDim(1);
-  // int64 in_channels = filter.getDim(2);
-  int64 out_channels = filter.getDim(3);
+void convMatrix::convolve(convMatrix &filter, int stride, bool samePadding,  convMatrix &target) {
+  int batch = getDim(0);
+  int in_height = getDim(1);
+  int in_width = getDim(2);
+  int in_channels = getDim(3);
+  int filter_height = filter.getDim(0);
+  int filter_width = filter.getDim(1);
+  // int in_channels = filter.getDim(2);
+  int out_channels = filter.getDim(3);
   assert(in_channels == filter.getDim(2)); 
-  for (int64 b = 0; b < batch; b++) {
-  	for (int64 i = 0; i*stride < in_height; i++) {
-      for (int64 j = 0; j*stride < in_width; j++) {
-        for (int64 k = 0; k < out_channels; k++){
+  for (int b = 0; b < batch; b++) {
+  	for (int i = 0; i*stride < in_height; i++) {
+      for (int j = 0; j*stride < in_width; j++) {
+        for (int k = 0; k < out_channels; k++){
           float pixVal = 0;
-          for (int64 c = 0; c < in_channels; c++) {
-            int64 startRow = i*stride - filter_height/2;
-            int64 startCol = j*stride - filter_width/2;
-            for (int64 di = 0; di < filter_height; di++) {
-              for (int64 dj = 0; dj < filter_width; dj++) {
-                int64 curRow = startRow + di;
-                int64 curCol = startCol + dj;
+          for (int c = 0; c < in_channels; c++) {
+            int startRow = i*stride - filter_height/2;
+            int startCol = j*stride - filter_width/2;
+            for (int di = 0; di < filter_height; di++) {
+              for (int dj = 0; dj < filter_width; dj++) {
+                int curRow = startRow + di;
+                int curCol = startCol + dj;
                 if (curRow > -1 && curRow < in_height && curCol > -1 && curCol < in_width) {
                   pixVal += (*this)(b, curRow, curCol, c) * filter(di, dj, c, k);
                 }
@@ -75,10 +75,10 @@ void convMatrix::convolve(convMatrix &filter, int64 stride, bool samePadding,  c
 }
 
 void convMatrix::flatten(Matrix &target) {
-  for (int64 i = 0; i < _D0; i++) {
-    for (int64 j = 0; j < _D1; j++) {
-      for (int64 m = 0; m < _D2; m++) {
-        for (int64 n = 0; n < _D3; n++) {
+  for (int i = 0; i < _D0; i++) {
+    for (int j = 0; j < _D1; j++) {
+      for (int m = 0; m < _D2; m++) {
+        for (int n = 0; n < _D3; n++) {
           target(i,j*_D2*_D3 + m*_D3 +n) = (*this)(i,j,m,n);
         }
       }
